pyproject.toml
[tool.poetry]
name = "gasPriceMessager"
version = "0.1.0"
description = ""
authors = ["Luke Pitstick <lukepitstick06@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
twilio = "^7.12.0"
python-dotenv = "^0.20.0"
beautifulsoup4 = "^4.11.1"
lxml = "^4.9.1"
cfscrape = "^2.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
